Fix: Fallback manual copy for update_localdb()
Created copytree() in utils.c that copies a directory recursively. This
is because the rename() will fail when the old and new names are on
different filesystems. update_localdb() will use copytree() if this
particular error occurs after trying to use rename() to move the database
from /tmp to $HOME. Also created copyfile() function as a helper function
for copytree().