Fix fish completion by escaping characters
Escape every tldr page name to prevent fish syntax errors.

This fixes the problem in which special characters in fish were not
escaped when fish shell completion was added in #17, because at the time
none of the pages in <https://github.com/tldr-pages/tldr> had names that
were syntactically special in fish. This meant that escaping was
unnecessary.

However, as a result of <https://github.com/tldr-pages/tldr/pull/7530>,
which added the files `[.md` and `[[.md`, the fish shell autocompletion
will now cause a syntax error in fish whenever autocompletion triggers.

This syntax error fills stderr with long error messages and prevents the
fish autocompletion from running. This problem is fixed by escaping
characters like `[`, which is done in this commit.

See also: <https://github.com/tldr-pages/tldr/pull/7530>, #17