Prevent files from displaying in fish autocomplete
When using `tldr` to show a tldr-page, specifying a file path instead of
a name of a page is not allowed and will simply fail with the message
`This page doesn't exist yet!`.

However, when using fish autocompletion, it will display file and
directory names along with the names of tldr-pages, even though using
those suggestions results in an error.

This commit tells fish that it should not complete filenames.

Documentation: <https://fishshell.com/docs/current/cmds/complete.html>