Remove redundant options in fish completion
The documentation for `complete` lists `-x` as an alias for `-rf`.

This allows for some options to be removed because their functionality
is already enabled by other flags.

> -x or --exclusive
>
>     Short for -r and -f.

Source: <https://fishshell.com/docs/current/cmds/complete.html>

This change is just refactoring. No functionality is different compared
to the last commit.